Add include serializer methods, move into mixin
