+++ app/serializers/directory_item_serializer.rb
@@ -1,27 +1,7 @@
 class DirectoryItemSerializer < ApplicationSerializer
 
   class UserSerializer < UserNameSerializer
-    attributes :primary_group_name,
-               :primary_group_flair_url,
-               :primary_group_flair_bg_color,
-               :primary_group_flair_color
-
-    def primary_group_name
-      return nil unless object&.primary_group_id
-      object&.primary_group&.name
-    end
-
-    def primary_group_flair_url
-      object&.primary_group&.flair_url
-    end
-
-    def primary_group_flair_bg_color
-      object&.primary_group&.flair_bg_color
-    end
-
-    def primary_group_flair_color
-      object&.primary_group&.flair_color
-    end
+    include UserPrimaryGroupMixin
   end
 
   attributes :id,
